Add files via upload
This project is a web application that allows users to download profile pictures from Instagram. It is built using the Flask web framework, integrating HTML, CSS, and JavaScript for the front-end. Users enter an Instagram username through a form on the website, and upon submission, the application fetches the profile picture using a Python script. The downloaded image is then presented to the user for download.

Key Features:

1. Flask web application framework for handling HTTP requests and responses.
2. User-friendly front-end with HTML form for inputting Instagram usernames.
3.Integration of CSS for styling and JavaScript for enhanced user interaction.
4.Utilizes Python scripts to scrape Instagram for profile pictures.
5.Profile pictures are downloaded and made available for the user to save.